+++ packages/fluentui/react-northstar/src/components/Tooltip/Tooltip.tsx
@@ -277,7 +277,10 @@ Tooltip.propTypes = {
   defaultOpen: PropTypes.bool,
   mountNode: customPropTypes.domNode,
   mouseLeaveDelay: PropTypes.number,
-  offset: PropTypes.string,
+  offset: PropTypes.oneOfType([
+    PropTypes.func,
+    PropTypes.arrayOf(PropTypes.number) as PropTypes.Requireable<[number, number]>,
+  ]),
   open: PropTypes.bool,
   onOpenChange: PropTypes.func,
   pointing: PropTypes.bool,
@@ -288,12 +291,14 @@ Tooltip.propTypes = {
   content: customPropTypes.shorthandAllowingChildren,
   unstable_pinned: PropTypes.bool,
   flipBoundary: PropTypes.oneOfType([
-    PropTypes.object as PropTypes.Requireable<Element>,
-    PropTypes.oneOf<'scrollParent' | 'window' | 'viewport'>(['scrollParent', 'window', 'viewport']),
+    PropTypes.object as PropTypes.Requireable<HTMLElement>,
+    PropTypes.arrayOf(PropTypes.object) as PropTypes.Requireable<HTMLElement[]>,
+    PropTypes.oneOf<'clippingParents' | 'window' | 'scrollParent'>(['clippingParents', 'window', 'scrollParent']),
   ]),
   overflowBoundary: PropTypes.oneOfType([
-    PropTypes.object as PropTypes.Requireable<Element>,
-    PropTypes.oneOf<'scrollParent' | 'window' | 'viewport'>(['scrollParent', 'window', 'viewport']),
+    PropTypes.object as PropTypes.Requireable<HTMLElement>,
+    PropTypes.arrayOf(PropTypes.object) as PropTypes.Requireable<HTMLElement[]>,
+    PropTypes.oneOf<'clippingParents' | 'window' | 'scrollParent'>(['clippingParents', 'window', 'scrollParent']),
   ]),
 };
 Tooltip.handledProps = Object.keys(Tooltip.propTypes) as any;